- Added -percentage option
git-svn-id: http://svn.nate.deepcreek.org.au/svn/tools/TSNowAndNext/trunk@19 c83a0d04-e8ae-47eb-aa4b-142be30de269

+++ functions.h
@@ -14,7 +14,7 @@ char* sprintToString(BYTE* pCurr, int count);
 void Mymemcpy(void *dest, const void*src, size_t length);
 void MyReadFile(HANDLE hFile, LPVOID lpBuffer, DWORD nNumberOfBytesToRead, LPDWORD lpNumberOfBytesRead, LPOVERLAPPED lpOverlapped);
 
-DWORD setFilePointer(HANDLE file, LONGLONG llDistanceToMove, DWORD dwMoveMethod);
+LONGLONG setFilePointer(HANDLE file, LONGLONG llDistanceToMove, DWORD dwMoveMethod);
 LONGLONG getFileSize(HANDLE file);
 
 #endif
